requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
bidict==0.23.1
build==1.2.1
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2024.2
cryptography==43.0.1
defusedxml==0.8.0rc2
dill==0.3.8
dj-rest-auth==4.0.0
Django==5.0
django-admin-interface==0.28.6
django-allauth==0.52.0
django-colorfield==0.11.0
django-crispy-forms==2.1
django-environ==0.11.2
django-filter==24.3
django-rest-auth==0.9.5
django-rest-authtoken==2.1.4
django-six==1.0.5
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.15.1
djangorestframework-api-key==2.0.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
dnspython==2.6.1
docutils==0.21.2
drf-spectacular==0.27.2
drf-test-generator==0.0.6
email_validator==2.2.0
factory_boy==3.3.1
Faker==30.8.0
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==7.2.0
inflection==0.3.1
iniconfig==2.0.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==25.2.1
lazy-object-proxy==1.10.0
lazy_loader==0.4
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
model-bakery==1.20.0
more-itertools==10.3.0
nh3==0.2.17
numpy==1.26.4
oauthlib==3.2.2
optional-django==0.3.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pipdeptree==2.16.2
pkginfo==1.11.1
platformdirs==4.2.2
pluggy==1.5.0
psutil==5.9.8
psycopg2==2.9.9
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.9.0
pyproject_hooks==1.1.0
pytest==7.4.4
pytest-assert-utils==0.3.1
pytest-common-subject==1.1.1
pytest-django==4.9.0
pytest-drf==1.1.3
pytest-factoryboy==2.7.0
pytest-fixture-order==0.1.4
pytest-lambda==1.3.0
pytest-reverse==1.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-slugify==8.0.4
python-socketio==5.11.3
python3-openid==3.2.0
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
react==4.3.0
readme_renderer==43.0
redis==5.0.6
referencing==0.35.1
reflex==0.5.4
reflex-hosting-cli==0.1.13
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rpds-py==0.20.0
setuptools==69.5.1
shellingham==1.5.4
simple-colors==0.1.5
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
SQLAlchemy==2.0.30
sqlmodel==0.0.19
sqlparse==0.5.0
starlette==0.37.2
starlette-admin==0.14.0
tabulate==0.9.0
text-unidecode==1.3
tomlkit==0.12.5
twine==5.1.0
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.24.0.post1
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
zipp==3.19.2
